Refactor: separate feature screen
