+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/AuthRoute.kt
@@ -1,0 +1,13 @@
+package com.teamoffroad.feature.auth
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun AuthRoute(
+    padding: PaddingValues,
+) {
+    AuthScreen(
+        padding = padding,
+    )
+}
