+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/ExploreRoute.kt
@@ -1,0 +1,13 @@
+package com.teamoffroad.feature.explore
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun ExploreRoute(
+    padding: PaddingValues,
+) {
+    ExploreScreen(
+        padding = padding,
+    )
+}
