+++ feature/home/src/main/java/com/teamoffroad/feature/home/HomeRoute.kt
@@ -1,0 +1,13 @@
+package com.teamoffroad.feature.home
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun HomeRoute(
+    padding: PaddingValues,
+) {
+    HomeScreen(
+        padding = padding,
+    )
+}
