+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/MypageRoute.kt
@@ -1,0 +1,13 @@
+package com.teamoffroad.feature.mypage
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun MypageRoute(
+    padding: PaddingValues,
+) {
+    MypageScreen(
+        padding = padding,
+    )
+}
